Move config to subdirectory
This will allow us to generate a sample config file when a user calls `vin advise`
