+++ main.go
@@ -7,6 +7,7 @@ import (
 	"github.com/grpc-ecosystem/go-grpc-middleware"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap"
 	"github.com/grpc-ecosystem/go-grpc-middleware/tags"
+	"github.com/vinyl-linux/vin/config"
 	"github.com/vinyl-linux/vin/server"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
@@ -50,7 +51,7 @@ func Setup() *grpc.Server {
 	sugar.Info("starting")
 	sugar.Info("loading config")
 
-	c, err := LoadConfig()
+	c, err := config.Load(configFile)
 	if err != nil {
 		sugar.Panic(err)
 	}
