+++ os_test.go
@@ -5,6 +5,8 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
+
+	"github.com/vinyl-linux/vin/config"
 )
 
 func TestUntar(t *testing.T) {
@@ -70,7 +72,7 @@ func TestExecute(t *testing.T) {
 				}
 			}()
 
-			err := execute(test.dir, test.command, output)
+			err := execute(test.dir, test.command, output, config.Config{})
 			close(output)
 			if err == nil && test.expectError {
 				t.Errorf("expected error")
